# front/openapi/openApi.public.yaml
openapi: 3.0.1
info:
  title: Bank API — public (no auth routes)
  description: >
    Все пути кроме `/user-service/auth/*`. Для user/admin и BFF без операций входа/регистрации.
    Полный контракт — openApi.backend-gateway.yaml; только auth — openApi.sso.yaml.
  version: "1.0"
servers:
  - url: http://localhost:8000/api
    description: >
      Base URL BFF (не шлюз бекенда). В проде подставьте origin своего BFF.
security:
  - BffSessionCookie: []
paths:
...
  /user-service/users/{id}/edit:
    put:
      tags:
        - Users
      operationId: editUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEditModelDto'
        required: true
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user-service/account/edit:
    put:
      tags:
        - Account
      operationId: editUser_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEditModelDto'
        required: true
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user-service/users:
    get:
      tags: 
        - Users
      operationId: getAllUsers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /user-service/users/{id}:
    get:
      tags:
        - Users
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Users
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user-service/users/{id}/isActive:
    get:
      tags:
        - Users
      operationId: isUserActiveById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /user-service/internal/users/{id}:
    get:
      tags:
        - Users Internal
      operationId: getUserById_1
      security:
        - BffServiceApiKey: []
      description: >
        Обычно вызывается только BFF/сервисом, не браузером. Аутентификация сервисным ключом;
        сессионная кука для этого пути не используется.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-API-KEY
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user-service/account:
    get:
      tags:
        - Account
      operationId: getUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  # Core API Endpoints
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    UserEditModelDto:
      required:
        - name
      type: object
      properties:
        name:
          type: string
        newRoles:
          description: >
            Новый набор ролей. Пустой массив снимает все роли (пользователь может остаться без ролей).
            Несколько ролей одновременно допустимы (например CLIENT и WORKER).
          uniqueItems: true
          type: array
          minItems: 0
          items:
            type: string
            enum:
            - CLIENT
            - WORKER
        active:
          type: boolean
          description: >
            При указании — выставить флаг активности пользователя (доступ / блокировка).
    UserDto:
      required:
        - email
        - id
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        roles:
          description: >
            Роли пользователя. Может быть пустым: у пользователя может не быть ни одной роли.
            Значения CLIENT и WORKER не взаимоисключающие — допустимы оба одновременно.
          uniqueItems: true
          type: array
          minItems: 0
          items:
            type: string
            enum:
              - CLIENT
              - WORKER
              - BLOCKED_CLIENT
              - BLOCKED_WORKER
        active:
          type: boolean
...
    BffErrorBody:
      type: object
      description: Обобщённое тело ошибки (BFF/шлюз; фактический JSON может отличаться).
      properties:
        message:
          type: string
        code:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    BadRequest:
      description: Некорректный запрос (синтаксис, параметры пути/запроса).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    Unauthorized:
      description: Нет или недействительная аутентификация (сессия, JWT, API-ключ).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    Forbidden:
      description: Доступ запрещён (недостаточно прав).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
...
    UnprocessableEntity:
      description: Ошибка валидации тела запроса или бизнес-правил.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    InternalServerError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
  securitySchemes:
    BffSessionCookie:
      type: apiKey
      in: cookie
      name: fins_session
      description: >
        Сессия браузера с BFF: httpOnly (и при необходимости Secure / SameSite) — имя куки задаёт BFF.
        Плейсхолдер `fins_session` замените на фактическое имя в своём BFF. Браузер не получает JWT бекенда.
    BffServiceApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Сервисный ключ BFF при вызове internal-маршрутов (не сессия пользователя).